learn-django/pyproject.toml
[project]
name = "learn-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=3.16.1",
]
